backend/requirements.txt
flask
flask-cors
psycopg2-binary
requests
python-dotenv
bcrypt
PyJWT
apscheduler
reportlab
Flask-Limiter